Rename memoryindex* -> memory_index*.
