+++ searchlib/src/tests/memoryindex/memory_index/.gitignore
@@ -1,0 +1,3 @@
+.depend
+Makefile
+searchlib_memory_index_test_app
