--- searchlib/src/tests/memoryindex/memoryindex/.gitignore
@@ -1,5 +1,0 @@
-.depend
-Makefile
-memoryindex_test
-sourceselectorwriter_test
-searchlib_memoryindex_test_app
