--- searchlib/src/tests/memoryindex/memoryindex/CMakeLists.txt
@@ -1,8 +1,0 @@
-# Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-vespa_add_executable(searchlib_memoryindex_test_app TEST
-    SOURCES
-    memoryindex_test.cpp
-    DEPENDS
-    searchlib
-)
-vespa_add_test(NAME searchlib_memoryindex_test_app COMMAND searchlib_memoryindex_test_app)
